requirements.txt
img2pdf==0.6.3
pillow==12.1.0
